backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
cryptography==41.0.7
pycryptodome==3.19.0
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
stripe==7.8.0
pyotp==2.9.0
qrcode==7.4.2
slowapi==0.1.9
apscheduler==3.10.4
requests==2.31.0
